backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
rq==1.15.1
pydantic[email]==2.5.0
google-search-results==2.4.2
openai==1.45.0
python-multipart==0.0.6

